requirements.txt
langchain
langchain-community
langchain-openai
openai
pydantic
duckduckgo-search
ddgs
pypdf
python-dotenv
requests
pytest
pytest-mock
pytest-asyncio